split layouts and finish icpdx hero
