+++ app/(legacy)/join-us/AccordionSection.tsx
