+++ app/(legacy)/messages/[id]/page.tsx
