+++ app/layout.tsx
@@ -23,15 +23,7 @@ export default function RootLayout({
       <body>
         <Providers>
           <Header />
-          <Box
-            sx={{
-              mb: 5,
-              px: '4vw',
-              minHeight: '90vh',
-            }}
-          >
-            {children}
-          </Box>
+          {children}
           <Footer />
         </Providers>
       </body>
