+++ components/Section.tsx
@@ -1,5 +1,7 @@
+import type { StaticImageData } from 'next/image'
+
 import { Box, Stack } from '@mui/material'
-import Image, { StaticImageData } from 'next/image'
+import Image from 'next/image'
 
 interface SectionProps {
   children: React.ReactElement
@@ -27,10 +29,9 @@ export function ElevatedSection({ children }: SectionProps) {
   )
 }
 
-export function ImageSection({ children, src, alt }: ImageSectionProps) {
+export function LegacyImageSection({ children, src, alt }: ImageSectionProps) {
   return (
     <Stack
-      component="section"
       sx={{
         py: '3.5vw',
         px: '3vw',
@@ -57,5 +58,15 @@ export function ImageSection({ children, src, alt }: ImageSectionProps) {
         {children}
       </Box>
     </Stack>
+  )
+}
+
+export function ImageSection({ children, src, alt }: ImageSectionProps) {
+  return (
+    <section className={'relative overflow-hidden'}>
+      <Image src={src} alt={alt} fill priority objectFit="cover" />
+      <div className="from-primary-dark-blue to-primary-light-blue absolute inset-0 bg-gradient-to-b opacity-50" />
+      <div className="relative">{children}</div>
+    </section>
   )
 }
